Add separate tests directory
