+++ tests/test_hsm.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-from hsm import HierachicalStateMachine, UnsupportedMessageType,\
+from ..other.hsm.hsm import HierachicalStateMachine, UnsupportedMessageType,\
     UnsupportedState, UnsupportedTransition, Active, Standby, Suspect, Failed
 from sys import version_info
